requirements.txt
streamlit>=1.25.0
pillow>=9.0.0

//how to install?
//1. open cmd
//2. pip install -r requirements.txt
